+++ stips/astro_image/astro_image.py
@@ -430,6 +430,9 @@ class AstroImage(object):
     @property
     def psf_constructor(self):
         import webbpsf
+        #**WFIRST_REMNANT**
+        if not hasattr(webbpsf, self.telescope.lower()) and self.telescope.lower() == 'roman':
+            return getattr(getattr(webbpsf, 'wfirst'), self.instrument)()
         return getattr(getattr(webbpsf, self.telescope), self.instrument)()
     
     
@@ -1625,7 +1628,7 @@ class AstroImage(object):
     
     
     INSTRUMENT_DEFAULT = {
-                            'telescope': 'wfirst',
+                            'telescope': 'roman',
                             'instrument': 'WFI',
                             'filter': 'F062',
                             'detector': {